IRIS-4208 | Name refactoring: global-notifications -> notification-types
