+++ front/main/app/utils/notification-types.js
